Break out basic matrix tests into seperate files
