+++ test/unit/math/gpu/opencl_add_test.cpp
@@ -1,6 +1,6 @@
 #ifdef STAN_OPENCL
 #include <stan/math/prim/mat.hpp>
-#include <stan/math/gpu/basic_matrix_gpu.hpp>
+#include <stan/math/gpu/add_opencl.hpp>
 #include <gtest/gtest.h>
