+++ test/unit/math/gpu/opencl_subtract_test.cpp
@@ -1,6 +1,6 @@
 #ifdef STAN_OPENCL
 #include <stan/math/prim/mat.hpp>
-#include <stan/math/gpu/basic_matrix_gpu.hpp>
+#include <stan/math/gpu/subtract_opencl.hpp>
 #include <gtest/gtest.h>
 
 TEST(MathMatrixGPU, subtract_v_exception_pass) {
