+++ test/unit/math/gpu/opencl_zeros_test.cpp
@@ -1,6 +1,6 @@
 #ifdef STAN_OPENCL
 #include <stan/math/prim/mat.hpp>
-#include <stan/math/gpu/basic_matrix_gpu.hpp>
+#include <stan/math/gpu/zeros_opencl.hpp>
 #include <gtest/gtest.h>
 #include <algorithm>
 
@@ -10,20 +10,13 @@ TEST(MathMatrixGPU, zero_m_exception_pass) {
   EXPECT_NO_THROW(stan::math::zeros(m));
   EXPECT_NO_THROW(stan::math::zeros(m, stan::math::UPPER));
   EXPECT_NO_THROW(stan::math::zeros(m, stan::math::LOWER));
-  
+
   stan::math::matrix_gpu m0;
   EXPECT_NO_THROW(stan::math::zeros(m0));
   EXPECT_NO_THROW(stan::math::zeros(m0, stan::math::LOWER));
   EXPECT_NO_THROW(stan::math::zeros(m0, stan::math::UPPER));
 }
 
-TEST(MathMatrixGPU, identity_m_exception_pass) {
-  stan::math::matrix_gpu m(1, 1);
-  stan::math::matrix_gpu m0;
-
-  EXPECT_NO_THROW(m = stan::math::identity(1));
-  EXPECT_NO_THROW(m0 = stan::math::identity(0));
-}
 
 TEST(MathMatrixGPU, zero_m_value_check) {
   stan::math::matrix_d m0(2, 2);
@@ -56,17 +49,4 @@ TEST(MathMatrixGPU, zero_m_value_check) {
   EXPECT_EQ(2, m0_dst(1, 1));
 }
 
-TEST(MathMatrixGPU, identity_m_value_check) {
-  stan::math::matrix_d m0(2, 2);
-  m0 << 2, 2, 2, 2;
-  stan::math::matrix_gpu m(m0);
-
-  EXPECT_NO_THROW(m = stan::math::identity(2));
-
-  stan::math::copy(m0, m);
-  EXPECT_EQ(1, m0(0, 0));
-  EXPECT_EQ(0, m0(0, 1));
-  EXPECT_EQ(0, m0(1, 0));
-  EXPECT_EQ(1, m0(1, 1));
-}
 #endif
